requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.8.5
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
altair==5.0.1
    # via streamlit
annotated-types==0.5.0
    # via pydantic
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
    #   streamlit
blinker==1.6.2
    # via streamlit
cachetools==5.3.1
    # via
    #   google-auth
    #   streamlit
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.0.4
    # via streamlit
dataclasses-json==0.5.14
    # via langchain
exceptiongroup==1.1.3
    # via pytest
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.10
    # via gitpython
gitpython==3.1.32
    # via streamlit
google-api-core==2.11.1
    # via langchain
google-auth==2.22.0
    # via google-api-core
googleapis-common-protos==1.60.0
    # via google-api-core
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via streamlit
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.19.0
    # via altair
jsonschema-specifications==2023.7.1
    # via jsonschema
langchain==0.0.272
    # via -r requirements.in
langsmith==0.0.26
    # via langchain
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.5
    # via langchain
numpy==1.25.2
    # via
    #   -r requirements.in
    #   altair
    #   langchain
    #   numexpr
    #   opencv-python
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==0.27.9
    # via -r requirements.in
opencv-python==4.8.0.76
    # via -r requirements.in
packaging==23.1
    # via
    #   marshmallow
    #   pytesseract
    #   pytest
    #   streamlit
pandas==2.0.3
    # via
    #   altair
    #   streamlit
pillow==10.0.0
    # via
    #   pytesseract
    #   streamlit
pluggy==1.2.0
    # via pytest
protobuf==4.24.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   streamlit
pyarrow==13.0.0
    # via streamlit
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.3.0
    # via
    #   langchain
    #   langsmith
pydantic-core==2.6.3
    # via pydantic
pydeck==0.8.1b0
    # via streamlit
pympler==1.0.1
    # via streamlit
pytesseract==0.3.10
    # via -r requirements.in
pytest==7.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via langchain
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   langchain
    #   langsmith
    #   openai
    #   streamlit
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
semver==3.0.1
    # via streamlit
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sqlalchemy==2.0.20
    # via langchain
streamlit==1.9.0
    # via -r requirements.in
tenacity==8.2.3
    # via langchain
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via pytest
toolz==0.12.0
    # via altair
tornado==6.3.3
    # via streamlit
tqdm==4.66.1
    # via openai
typing-extensions==4.7.1
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
tzlocal==5.0.1
    # via streamlit
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
validators==0.21.2
    # via streamlit
watchdog==3.0.0
    # via streamlit
yarl==1.9.2
    # via aiohttp
zipp==3.16.2
    # via importlib-metadata
